% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODproNIm.R
\name{ODproNIm}
\alias{ODproNIm}
\title{Delete outlier cases}
\usage{
ODproNIm(data.df, imp = "mice", missing_ratio = 0.1,
  cut_off_ratio = 0.2, iter = 1000)
}
\arguments{
\item{data.df}{target dataset}

\item{imp}{imputation method name (mice (default) or missForest)}

\item{missing_ratio}{ratio to product NA}

\item{cut_off_ratio}{ratio of delete cases in whole data}

\item{iter}{times of iteration}
}
\description{
Delete outlier cases
}
\examples{
ODproNIm(iris[,-ncol(iris)],iter=10)

}
