% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OEdii.R
\name{OEdii}
\alias{OEdii}
\title{Eliminate outliers using deletion-imputation Iteration.}
\usage{
OEdii(
  data.df,
  imp = "mice",
  del_rate = 0.1,
  elim_rate = 0.2,
  iter = 1000,
  penl = "SQD"
)
}
\arguments{
\item{data.df}{outlier elimination target dataset}

\item{imp}{imputation method's name ('mice' (default setting) or 'missForest')}

\item{del_rate}{deletion rate to complete dataset}

\item{elim_rate}{outlier elimination rate}

\item{iter}{the number of of iteration}

\item{penl}{calculation method of residual}
}
\value{
$elim returns dataframe without outlier (eliminated)

$diff returns means of difference between imputed value and actual one about each field

$mcont returns the number of missing count of each field
}
\description{
Eliminate outliers using deletion-imputation Iteration.
}
\examples{
OEdii(iris[,-ncol(iris)], iter=100)
out <- OEdii(iris[,-ncol(iris)],iter=100)
out$elim

}
