% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OEprNIm.R
\name{OEprNIm}
\alias{OEprNIm}
\title{Eliminate outlier cases}
\usage{
OEprNIm(data.df, imp = "mice", na_ratio = 0.1, co_ratio = 0.2,
  iter = 1000, calc_res = "arm")
}
\arguments{
\item{data.df}{target dataset}

\item{imp}{imputation method name (mice (default) or missForest)}

\item{na_ratio}{ratio of producting NA}

\item{co_ratio}{ratio of delete cases in whole data}

\item{iter}{the number of of iteration}

\item{calc_res}{calculation method of residual}
}
\description{
Eliminate outlier cases
}
\examples{
OEprNIm(iris[,-ncol(iris)],iter=10)

}
