% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OEprNIm.R
\name{OEprNIm}
\alias{OEprNIm}
\title{Eliminate outlier cases by using generating NA and imputation chain.}
\usage{
OEprNIm(data.df, imp = "mice", na_ratio = 0.1, co_ratio = 0.2,
  iter = 1000, calc_res = "arm")
}
\arguments{
\item{data.df}{target dataset}

\item{imp}{imputation method name ('mice' (default setting) or 'missForest')}

\item{na_ratio}{ratio of generating NA}

\item{co_ratio}{ratio of eliminated cases by whole data}

\item{iter}{the number of of iteration}

\item{calc_res}{calculation method of residual}
}
\value{
$elim returns eliminated dataframe, $diff returns difference of each field and $mcont returns the number of missing count of each field
}
\description{
Eliminate outlier cases by using generating NA and imputation chain.
}
\examples{
OEprNIm(iris[,-ncol(iris)],iter=10)

}
